Split automation event source ownership
